+++ src/__tests__/babelTiming.test.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 const util = require('util');
 const exec = util.promisify(require('child_process').exec);
 const rimraf = require('rimraf');
-const {babelTiming} = require('../index');
+const { babelTiming } = require('../index');
 const FIXTURES = '__fixtures__';
 const {
   expectedResults,
@@ -116,26 +116,6 @@ describe('babelTiming', () => {
           ])
         );
       });
-    });
-  });
-
-  describe('"--read-results" CLI option', () => {
-    it('render results object provided as path', async () => {
-      const resultsPath = path.join(FIXTURES, 'results.json');
-      const expectedResultsPath = path.join(
-        FIXTURES,
-        'generated.results.temp.json'
-      );
-
-      await exec(
-        `node cli.js --read-results ${resultsPath} --output-path ${expectedResultsPath} --output json`
-      );
-
-      const expected = JSON.parse(fs.readFileSync(resultsPath));
-      const actual = JSON.parse(fs.readFileSync(expectedResultsPath));
-      rimraf.sync(expectedResultsPath);
-
-      expect(actual).toEqual(expected);
     });
   });