+++ src/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  babelTiming: require('./standalone'),
-  Timer: require('./Timer'),
-  timersCollection: require('./timersCollection'),
-  render: require('./render'),
-};