+++ src/index.ts
@@ -1,0 +1,5 @@
+import * as timersCollection from './timersCollection';
+export { timersCollection };
+export { default as babelTiming } from './standalone';
+export { default as Timer } from './Timer';
+export { default as render } from './render';