+++ src/render/cliRenderer/index.ts
@@ -1,8 +1,19 @@
-var differ = require('ansi-diff-stream');
-const Table = require('./Table');
-const {enableKeyPressEvent} = require('./utils');
+import differ from 'ansi-diff-stream';
+import Table from './Table';
+import { enableKeyPressEvent } from './utils';
+import { ResultByFile, ResultByPlugin } from '../../types';
 
-function renderFileList({results, selected = 0, diff, paginationSize} = {}) {
+function renderFileList({
+  results,
+  selected,
+  diff,
+  paginationSize,
+}: {
+  results: ResultByFile[];
+  selected: number;
+  diff: ReturnType<typeof differ>;
+  paginationSize: number;
+}) {
   const output = new Table({
     title: 'Babel timing - trasformed files',
     entries: results,
@@ -15,7 +26,12 @@ function renderFileList({results, selected = 0, diff, paginationSize} = {}) {
     onSelected: selected => {
       diff.clear();
       output.unmount();
-      renderFileDetails({results, resultIndex: selected, diff, paginationSize});
+      renderFileDetails({
+        results,
+        resultIndex: selected,
+        diff,
+        paginationSize,
+      });
     },
     onSelectedCommandInfo: 'show file detail',
     paginationSize,
@@ -25,7 +41,17 @@ function renderFileList({results, selected = 0, diff, paginationSize} = {}) {
   });
 }
 
-function renderFileDetails({results, resultIndex, diff, paginationSize} = {}) {
+function renderFileDetails({
+  results,
+  resultIndex,
+  diff,
+  paginationSize,
+}: {
+  results: ResultByFile[];
+  resultIndex: number;
+  diff: ReturnType<typeof differ>;
+  paginationSize: number;
+}) {
   const fileResult = results[resultIndex];
   const output = new Table({
     title: `Babel timing - info for file: ${fileResult.name}`,
@@ -39,7 +65,7 @@ function renderFileDetails({results, resultIndex, diff, paginationSize} = {}) {
     onEscape: () => {
       diff.clear();
       output.unmount();
-      renderFileList({results, selected: resultIndex, diff, paginationSize});
+      renderFileList({ results, selected: resultIndex, diff, paginationSize });
     },
     onEscapeCommandInfo: 'back to results list',
     paginationSize,
@@ -49,7 +75,17 @@ function renderFileDetails({results, resultIndex, diff, paginationSize} = {}) {
   });
 }
 
-function renderPluginList({results, selected = 0, diff, paginationSize} = {}) {
+function renderPluginList({
+  results,
+  selected = 0,
+  diff,
+  paginationSize,
+}: {
+  results: ResultByPlugin[];
+  selected: number;
+  diff: ReturnType<typeof differ>;
+  paginationSize: number;
+}) {
   const output = new Table({
     title: 'Babel timing - plugins called',
     entries: results,
@@ -82,7 +118,12 @@ function renderPluginDetails({
   resultIndex,
   diff,
   paginationSize,
-} = {}) {
+}: {
+  results: ResultByPlugin[];
+  resultIndex: number;
+  diff: ReturnType<typeof differ>;
+  paginationSize: number;
+}) {
   const pluginResult = results[resultIndex];
   const output = new Table({
     title: `Babel timing - info for plugin: ${pluginResult.name}`,
@@ -96,7 +137,12 @@ function renderPluginDetails({
     onEscape: () => {
       diff.clear();
       output.unmount();
-      renderPluginList({results, selected: resultIndex, diff, paginationSize});
+      renderPluginList({
+        results,
+        selected: resultIndex,
+        diff,
+        paginationSize,
+      });
     },
     onEscapeCommandInfo: 'back to results list',
     paginationSize,
@@ -106,22 +152,37 @@ function renderPluginDetails({
   });
 }
 
-function renderer(results = [], {paginationSize} = {}) {
-  // Duck type results to tell if data is aggregated by files or plugins.
-  // @TODO Find a better way to adjust renderer on data type
-  isFileList = results[0].hasOwnProperty('plugins');
+// Duck type results to tell if data is aggregated by files or plugins.
+// @TODO Find a better way to adjust renderer on data type
+function isResultsByFile(
+  results: ResultByFile[] | ResultByPlugin[]
+): results is ResultByFile[] {
+  return results[0].hasOwnProperty('plugins');
+}
+
+export default function renderer(
+  results: ResultByFile[] | ResultByPlugin[] = [],
+  { paginationSize }: { paginationSize: number }
+) {
   enableKeyPressEvent();
 
   // Init ansi-diff-stream
   const diff = differ();
   diff.pipe(process.stdout);
 
-  const renderer = isFileList ? renderFileList : renderPluginList;
-  renderer({
-    results,
-    diff,
-    paginationSize,
-  });
+  if (isResultsByFile(results)) {
+    renderFileList({
+      results,
+      selected: 0,
+      diff,
+      paginationSize,
+    });
+  } else {
+    renderPluginList({
+      results,
+      selected: 0,
+      diff,
+      paginationSize,
+    });
+  }
 }
-
-module.exports = renderer;