+++ src/types/ansi-diff-stream/index.d.ts
@@ -1,0 +1,7 @@
+declare module 'ansi-diff-stream' {
+  export default function differ(): {
+    pipe: (stdout: NodeJS.WriteStream) => void;
+    clear: () => void;
+    write: (output: string) => void;
+  };
+}